ref : - update logic for get all user api - add create absent permission functionallity - add get all permission by user

+++ app/Http/Controllers/Api/UserController.php
@@ -5,14 +5,18 @@ namespace App\Http\Controllers\Api;
 use App\Models\User;
 use Illuminate\Http\Request;
 use App\Http\Controllers\Controller;
+use App\Models\AbsentPermission;
 use App\Models\Attende;
 use App\Models\AttendeCode;
+use App\Transformers\AbsentPermissionTransformer;
 use App\Transformers\AllUserTransformers;
 use App\Transformers\AttendeTransformers;
 use App\Transformers\Serializers\CustomSerializer;
 use App\Transformers\UserTransformer;
 use Carbon\Carbon;
+use Illuminate\Support\Facades\File;
 use Illuminate\Support\Facades\Hash;
+use Illuminate\Support\Facades\Storage;
 use Illuminate\Support\Facades\Validator;
 
 class UserController extends Controller
@@ -26,10 +30,11 @@ class UserController extends Controller
     public function index(Request $request)
     {
         $users = User::with(['gender', 'departemen'])
-            ->where('id', '!=', $request->user()->id)
-            ->pns()
-            ->orWhere
-            ->honorer()
+            ->where('id', '!=', $request->user()->id)->where(function ($query) {
+                return $query->pns()
+                    ->orWhere
+                    ->honorer();
+            })
             ->get();
         $users = fractal()->collection($users)
             ->transformWith(new AllUserTransformers)
@@ -38,24 +43,28 @@ class UserController extends Controller
         return setJson(true, 'Berhasil mengambil seluruh data pegawai!', $users, 200, []);
     }
 
+
     public function login(Request $request)
     {
         $validator = Validator::make($request->all(), [
             'phone' => 'required',
             'password' => 'required',
             'device_name' => 'required'
+        ], [
+            'required' => ':attribute tidak boleh kosong!'
         ]);
 
         if ($validator->fails()) {
             return setJson(false, 'Gagal masuk ke aplikasi!', [], 400, $validator->errors()->toArray());
         }
 
-        $user = User::where('phone', $request->phone)->with(['presensi', 'izin', 'departemen', 'gender'])->first();
+
+        $user = User::where('phone', wordwrap($request->phone, 4, " ", true))->with(['presensi', 'izin', 'departemen', 'gender'])->first();
 
         if (!$user || !Hash::check($request->password, $user->password)) {
-            return setJson(false, 'No handphone/password salah!', [], 400, ['message' => 'Data not found']);
+            return setJson(false, 'No handphone/password salah!', [], 400, ['message' => ['Data not found']]);
         }
-        $nextPresence  = $this->checkNextPresence($request, $user->id);
+        $nextPresence  = $this->checkNextPresence($user->id);
         $token = $user->createToken($request->device_name)->plainTextToken;
         $user = fractal()->item($user)
             ->transformWith(new UserTransformer($token, $nextPresence))
@@ -76,7 +85,9 @@ class UserController extends Controller
             'code' => 'required',
             'photo' => 'required',
             'latitude' => 'required',
-            'longitude' => 'required'
+            'longitude' => 'required',
+            'address' => 'required',
+            'file_name' => 'required',
         ]);
 
         if ($validator->fails()) {
@@ -132,12 +143,19 @@ class UserController extends Controller
             }
         }
 
+
+        $realImage = base64_decode($request->photo);
+        $imageName = $request->user()->name . "-" . $request->file_name;
+
+        Storage::disk('public')->put("presensi/" . now()->format('d_m_Y') . "/" . $code->tipe->name . "/$imageName",  $realImage);
+
         $update = $attende->update([
             'attend_time' => now(),
             'attende_status_id' => $status,
-            'photo' => $request->photo,
+            'photo' => "presensi/" . now()->format('d_m_Y') . "/" . $code->tipe->name . "/$imageName",
             'latitude' => $request->latitude,
-            'longitude' => $request->longitude
+            'longitude' => $request->longitude,
+            'address' => $request->address
         ]);
 
         if ($update) {
@@ -153,6 +171,71 @@ class UserController extends Controller
         return setJson(false, 'Kesalahan tidak diketahui!', [], 400, ['message' => 'Kesalahan tidak diketahui!']);
     }
 
+    public function createPermission(Request $request)
+    {
+        $validator = Validator::make(
+            $request->all(),
+            [
+                'title' => 'required',
+                'description' => 'required',
+                'photo' => 'required',
+                'due_date' => 'required',
+                'start_date' => 'required',
+                'file_name' => 'required',
+            ],
+            [
+                'required' => ':attribute tidak boleh kosong!'
+            ]
+        );
+
+        if ($validator->fails()) {
+            return setJson(false, 'Gagal', [], 400, $validator->errors());
+        }
+
+        if (Carbon::parse($request->due_date) <= now()) {
+            return setJson(false, 'Gagal', [], 400, ['tanggal_kadaluarsa' => ['Hanya boleh menambahkan surat izin tertanggal setelah hari ini.']]);
+        }
+
+        $realImage = base64_decode($request->photo);
+        $imageName = $request->title . "-" . now()->translatedFormat('l, d m Y') . "-" . $request->file_name;
+
+        Storage::disk('public')->put("izin/" . $request->user()->name . "/"   . $imageName,  $realImage);
+
+        $permission = AbsentPermission::create([
+            'user_id' => $request->user()->id,
+            'title' => $request->title,
+            'description' => $request->description,
+            'photo' => "izin/" . $request->user()->name . "/"   . $imageName,
+            'due_date' => Carbon::parse($request->due_date),
+            'start_date' => Carbon::parse($request->start_date),
+            'is_approved' => false
+        ]);
+
+        if ($permission) {
+            return setJson(
+                true,
+                'Berhasil',
+                fractal()->item($permission)
+                    ->transformWith(new AbsentPermissionTransformer)
+                    ->serializeWith(new CustomSerializer)->toArray(),
+                201,
+                []
+            );
+        }
+
+        return setJson(false, 'Gagal', [], 400, ['message' => ['Kesalahan tidak diketahui!']]);
+    }
+
+    public function permission(Request $request)
+    {
+        $permissions = AbsentPermission::where('user_id', $request->user()->id)->get();
+        $permissions = fractal()
+            ->collection($permissions, new AbsentPermissionTransformer)
+            ->serializeWith(new CustomSerializer)->toArray();
+
+        return setJson(true, 'Berhasil', $permissions, 200, []);
+    }
+
     /**
      * Display the specified resource.
      *
@@ -161,7 +244,7 @@ class UserController extends Controller
      */
     public function show(Request $request)
     {
-        $nextPresence  = $this->checkNextPresence($request, $request->user()->id);
+        $nextPresence  = $this->checkNextPresence($request->user()->id);
         $user = fractal()->item($request->user())
             ->transformWith(new UserTransformer($request->bearerToken(), $nextPresence))
             ->serializeWith(new CustomSerializer)->toArray();
@@ -196,7 +279,7 @@ class UserController extends Controller
             return setJson(false, 'Password salah!', [], 400, ['old_pass' => ['Password lama salah!']]);
         }
 
-        $nextPresence  = $this->checkNextPresence($request, $request->user()->id);
+        $nextPresence  = $this->checkNextPresence($request->user()->id);
 
         if ($request->user()->update(['password' => Hash::make($request->new_pass)])) {
             $user = fractal()->item($request->user())
@@ -216,12 +299,12 @@ class UserController extends Controller
      */
     public function logout(Request $request)
     {
-        $request->user()->tokens()->delete();
+        $request->user()->tokens()->where('id',  $request->user()->currentAccessToken()->id)->delete();
 
         return setJson(true, 'Berhasil keluar dari aplikasi!', [], 200);
     }
 
-    private function checkNextPresence($request, $userId)
+    private function checkNextPresence($userId)
     {
         $attendeCode = AttendeCode::with(['tipe'])->whereDate('created_at', today())->get();
 
@@ -243,6 +326,14 @@ class UserController extends Controller
                     $nextPresence = Attende::where(
                         [
                             ['attende_code_id', $attendeCode[$i]->id],
+                            ['user_id', $userId]
+                        ]
+                    )->first();
+                    break;
+                } else if (now()->format('H') < 9) {
+                    $nextPresence = Attende::where(
+                        [
+                            ['attende_code_id', $attendeCode[0]->id],
                             ['user_id', $userId]
                         ]
                     )->first();
